requirements.txt
crewai==0.28.0
crewai-tools==0.2.6
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-openai==0.0.5
langchain-community==0.0.10
duckduckgo-search==4.1.1
weaviate-client==3.26.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi==0.108.0
uvicorn==0.25.0
prometheus-client==0.19.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
scipy==1.11.4
